Split trampoline classes into core.h and stages.h
